Swap to CFFI 1.0 build style.
